# public/api.yaml
info:
  title: Coronatime API
  description: Here goes description
  version: 0.1.9
servers:
  - url: http://localhost:3000/wane/v1
    description: local server
  - url: http://producation.app/api
    description: production server
openapi: 3.0.0
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nom:
          type: string
          example: "Doe"
        prenom:
          type: string
          example: "John"
        login:
          type: string
          example: "johndoe"
        password:
          type: string
          format: password
          example: "Password123!"
        photo:
          type: string
          format: uri
          example: "https://example.com/path/to/photo.jpg"
      required:
        - nom
        - prenom
        - login
        - password
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      summary: Get a list of users
      tags:
        - Users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '401':
          description: Unauthenticated
      security:
        - bearerAuth: []
  /users/{id}:
    get:


      summary: Get a user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the use
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '401':
          description: Unauthenticated


    delete:
      summary: Delete a user
      tags:
        - Users
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '401':
          description: Unauthenticated


    put:
      summary: Update a user
      tags:
        - Users
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '401':
          description: Unauthenticated
...